central-server/requirements.txt
fastapi
uvicorn
pymongo
motor
beanie
passlib
python-jose
email-validator
python-multipart
types-requests
types-orjson
nats-py